requirements.txt
# Django
Django==4.2.7
djangorestframework==3.14.0

# Channels (WebSocket support)
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2

# Database (Python 3.13 Compatible)
psycopg[binary]==3.2.12

# File handling (Python 3.13 Compatible)
Pillow==11.0.0

# Authentication
djangorestframework-simplejwt==5.3.1

# CORS
django-cors-headers==4.3.1

# Environment variables
python-dotenv==1.0.1

# Redis client
redis==5.1.1

# Production server
gunicorn==21.2.0
whitenoise==6.6.0